checando se usuario tem permissao ao deletar projeto

+++ src/modules/project/use-cases/DeleteProjectUseCase.ts
@@ -2,15 +2,19 @@ import {
   NotParticipantInProjectError,
   ProjectNotFoundError,
 } from "@shared/use-cases/errors";
+import { RoleInsufficientPermissionError } from "@shared/use-cases/errors/RoleInsufficientPermissionError";
 import {
   INotParticipantInProjectErrorLanguage,
   IProjectNotFoundErrorLanguage,
+  IRoleInsufficientPermissionErrorLanguage,
 } from "@shared/use-cases/interfaces/languages";
 import {
   IDoesParticipantExistRepository,
   IDoesProjectExistRepository,
   IFindParticipantRoleInProjectRepository,
 } from "@shared/use-cases/interfaces/repositories";
+import { IInvalidRoleNameErrorLanguage } from "../entities/interfaces/languages";
+import { Role } from "../entities/value-objects";
 import { DeleteProjectDTO } from "./DTOs/DeleteProjectDTO";
 import { IDeleteProjectRepository } from "./interfaces/repositories";
 
@@ -21,7 +25,9 @@ export class DeleteProjectUseCase {
     private readonly doesParticipantExistRepository: IDoesParticipantExistRepository,
     private readonly findParticipantRoleInProjectRepository: IFindParticipantRoleInProjectRepository,
     private readonly projectNotFoundErrorLanguage: IProjectNotFoundErrorLanguage,
-    private readonly notParticipantInProjectErrorLanguage: INotParticipantInProjectErrorLanguage
+    private readonly notParticipantInProjectErrorLanguage: INotParticipantInProjectErrorLanguage,
+    private readonly invalidRoleNameErrorLanguage: IInvalidRoleNameErrorLanguage,
+    private readonly roleInsufficientPermissionErrorLanguage: IRoleInsufficientPermissionErrorLanguage
   ) {}
 
   async delete({
@@ -45,6 +51,21 @@ export class DeleteProjectUseCase {
       throw new NotParticipantInProjectError(
         accountEmailMakingRequest,
         this.notParticipantInProjectErrorLanguage
+      );
+
+    const participantRoleName =
+      await this.findParticipantRoleInProjectRepository.findParticipantRole({
+        projectId,
+        accountEmail: accountEmailMakingRequest,
+      });
+    const role = new Role(
+      participantRoleName,
+      this.invalidRoleNameErrorLanguage
+    );
+    if (!role.can("DELETE_PROJECT"))
+      throw new RoleInsufficientPermissionError(
+        participantRoleName,
+        this.roleInsufficientPermissionErrorLanguage
       );
 
     await this.deleteProjectRepository.deleteProject(projectId);
